feat(engine): align skill modifiers across engine and client
