+++ packages/engine/src/skills/generation/graph.ts
