+++ packages/engine/src/skills/generation/utils.ts
